pyproject.toml
[tool.poetry]
name = "gitcha-action"
version = "0.1.0"
description = ""
authors = ["shry <shry@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "gitcha"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.146"
openai = "^0.27.4"
unstructured = {extras = ["local-inference"], version = "^0.5.12"}
tiktoken = "^0.3.3"
pygithub = "^1.58.1"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
pyyaml = "^6.0"
python-frontmatter = "^1.0.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
pre-commit = "^3.2.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
